[Bugfix] Handle deeply-nested dependent params (#1661)
* Repro issue #1659

* [Fix] Handle deeply-nested dependencies with `given`.

Behind the scenes, each call to `requires` or other params DSL method
pushes an entry onto a flat list of validators. The nesting structure
that your parameters can take on is tracked as an up-tree separately
on each scope, but that relationship isn't used to traverse the validations.
So, when I moved the dependency checking out of `should_validate?` and into
the actual validation, the `given` dependency stopped taking effect after
you nested parameters more than one level deep underneath.

To restore the behavior, I made the validation check recursively upwards
to see if it should or should not validate that scope.

* Add changelog entry.